Transfer all the api view methods to their own file
